+++ packages/firebase-dart/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added `measurementId` and `appId` to `FirebaseOptions` (both required for analytics)
 - Added analytics and performance interop
 - Added `User.getIdTokenResult`.
+- Removed long-deprecated `Auth.fetchProvidersForEmail` function.
 - Updated documented JS library from `6.6.1` of the JS API to `7.4.0`.
 
 ## 6.0.0